==> 0a.directory_walk.ipynb <==
# Extract transcripts from .json files

import pandas as pd
import numpy as np
import json
import os

# Define diretory name
directory = "episode_extracts_walk"

# initiate df and columns
transcripts_dataframe = pd.DataFrame()
all_showid = []
all_eposideid = []
all_transcripts = []
all_confidence = []
faulty_files = 0 # variable for checking how many faulty files

# iterate over files in dir
for dirpath, dirnames, files in os.walk(directory):
    for filename in files:
        transcripts = []
        confidence = []
        try:  # handle problem with files in wrong format (.DS_Store)
            with open(os.path.join(dirpath, filename)) as file:
                file_content = json.load(file)
                
                # index .json and loop at correct indent
                results = file_content["results"]
                for i in range(len(results)):
                    # lists are named after their key in the preceeding dict, dicts are named after their order
                    first_dict = results[i]
                    alternatives = first_dict["alternatives"]
                    second_dict = alternatives[0] 
                    if len(second_dict) == 3: # handle problem with empty alternative lists
                        transcripts.append(second_dict["transcript"])
                        confidence.append(second_dict["confidence"])
            
            all_showid.append(dirpath[-6:]) # OBS: change to [-27:] when running in Ucloud
            all_eposideid.append(str(filename))
            all_transcripts.append("".join(transcripts))
            all_confidence.append(np.mean(confidence))  # weighted confidence?
        
        except:
            faulty_files += 1

# Print errors message
print("Number of files with wrong format: ", faulty_files)

# Create dataframe
transcripts_dataframe["show_id"] = all_showid
transcripts_dataframe["episode_id"] = all_eposideid
transcripts_dataframe["transcript"] = all_transcripts
transcripts_dataframe["avg_confidence"] = all_confidence # should we do a weighted confidence?
transcripts_dataframe["word_count"] = transcripts_dataframe.transcript.apply(lambda x: len(x.split(" ")))

# save new df as compressed csv
transcripts_dataframe.to_csv("transcripts_dataset.csv.gz", na_rep="Na", encoding="utf-8", compression="gzip", chunksize=1000)
# print result
print(transcripts_dataframe.shape)
transcripts_dataframe.head()
