==> 0b.data_wrangling.ipynb <==
# File for creating more features - not used in final version

import pandas as pd
import numpy as np

# NLP libs
import spacy, nltk, gensim, sklearn
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

# meta_data = pd.read_csv("metadata.tsv", sep = "/t")
# meta_data.columns

features = ["show_name", "show_description", "publisher", "language", "episode_name", "episode_description", "duration"]
features_df = meta_data[features]
features_df = features_df.dropna()  # drops 207 rows 

# # Remove newlines (initial analysis showed that eps descriptions do not contain newlines)
# episode_description_new = [" ".join(desc.split()) for desc in features_df.episode_description]
# features_df["episode_description_clean"] = episode_description_new # append df

### OBS: TAKES A LOT OF TIME! Approx. 60 min
# Make spacy objects according to English language model
nlp = spacy.load("en_core_web_sm")

# spacy_obj = [nlp(i) for i in features_df.episode_description]
features_df["eps_nlp"] = spacy_obj
# spacy_obj_show = [nlp(i) for i in features_df.show_description]
features_df["show_nlp"] = spacy_obj_show

# Tokenise episodes
eps_tokens_all = []
for episode in features_df.eps_nlp:
    eps_tokens_all.append([tok.text for tok in episode])

features_df["episode_tokens"] = eps_tokens_all

print(features_df.eps_nlp[0])
print(features_df.episode_tokens[0])

# Tokenise show
eps_tokens_all_show = []
for show in features_df.eps_nlp:
    eps_tokens_all_show.append([tok.text for tok in show])

features_df["show_tokens"] = eps_tokens_all_show

print(features_df.show_nlp[0])
print(features_df.show_tokens[0])
features_df.head()

# Add word count features
# Epsiode desc
eps_word_count = []
for desc in features_df.episode_tokens:
    eps_word_count.append(len(desc))

features_df["eps_word_count"] = eps_word_count

# Show desc
show_word_count = []
for desc in features_df.episode_tokens:
    show_word_count.append(len(desc))

features_df["show_word_count"] = show_word_count

# Sentiment analysis
sent_analyser = SentimentIntensityAnalyzer()

def sentiment(text, choice):
    scores = sent_analyser.polarity_scores(text)
    return scores[choice]

features_nlp["positive_score_eps"] = features_nlp.episode_description.apply(lambda x: sentiment(x, "pos"))
features_nlp["negative_score_eps"] = features_nlp.episode_description.apply(lambda x: sentiment(x, "neg"))
features_nlp["neutral_score_eps"] = features_nlp.episode_description.apply(lambda x: sentiment(x, "neu"))

**Up until this point, the metadata_features dataset is constructed**

features_nlp = pd.read_csv('metadata_features.csv.gz', sep = ",", compression='gzip')

# Write to file to avoid long processing in the future
# features_nlp.to_csv("metadata_features.csv.gz", index=False, compression='gzip', encoding='utf-8')
