==> 0c.RSS_feed.ipynb <==
# Access RSS feeds and retrive more features

import pandas as pd
import numpy as np
import feedparser
from datetime import datetime
from dateutil import parser
import time

metadata = pd.read_csv('/Users/ramon.habtezghi/Desktop/THESIS/REPO/Thesis/metadata.csv.gz', compression='gzip')
metadata.isna().sum()

metadata_err = metadata[metadata.category.isna()]
metadata_err.shape

# Loop through each row in the metadata DataFrame
for index, row in metadata_err.iterrows():
    # Extract the episode_name and episode_uri from the current row
    episode_name = row['episode_name'].strip()
    rss_link = row['rss_link']

    # Try to fetch the RSS feed from the episode_uri, and retry up to 3 times if there's an error
    for attempt in range(5):
        try:
            feed = feedparser.parse(rss_link)
            break
        except Exception as e:
            print(f"Error fetching RSS feed for {episode_name}: {e}")
            print(f"Retrying in 20 seconds... (Attempt {attempt+1}/5)")
            time.sleep(20)

    # If the maximum number of retries is reached, skip to the next row
    else:
        print(f"Max number of retries reached for {episode_name}. Skipping...")
        continue

    # Extract the category from the RSS feed's <channel> element
    try:
        channel = feed.feed
        category = channel.tags[0].get('term')
        # Update the current row with the fetched data
        metadata_err.at[index, 'category'] = category
        # print(category)
    except AttributeError:
        category = None
        continue

    # Loop through each item in the RSS feed
    for item in feed.entries:
        # Check if the current item's title matches the episode_name
        if item.title == episode_name:
            # Extract pubdate from the current item and convert to YYYY-MM-DD format
            pubdate = parser.parse(item.published).date()
            # print(pubdate)

            # Update the current row with the fetched data
            metadata_err.at[index, 'pubdate'] = pubdate            

metadata_err.isna().sum()

print(metadata_err.shape)
print(metadata.shape)

list(metadata_err.columns)

# merge the two dataframes based on their common column (id)
metadata_merged = pd.merge(metadata, metadata_err, on=
                           ['show_uri','show_name','show_description','publisher', 'language','rss_link','episode_uri','episode_name',
                            'episode_description','duration','show_filename_prefix','episode_filename_prefix'], how='left')

print(metadata_merged.columns)
metadata_merged.isna().sum()

# fill the missing values in the merged dataframe with the values from the metadata_err dataframe
metadata_merged['category_x'].fillna(metadata_merged['category_y'], inplace=True)
metadata_merged['pubdate_x'].fillna(metadata_merged['pubdate_y'], inplace=True)

# drop the duplicate columns
metadata_merged.drop(['category_y', 'pubdate_y'], axis=1, inplace=True)
metadata_merged.rename(columns={'category_x': 'category', 'pubdate_x': 'pubdate'}, inplace=True)

# the merged dataframe now contains the missing values from the metadata_err dataframe
metadata_merged.isna().sum()

metadata_merged.shape

# Save new metadata to file
metadata_merged.to_csv('metadata.csv.gz', compression='gzip', index=False)
