==> 0d.make_final_dataset.ipynb <==
# Join all data from previous steps to one file

import pandas as pd
import numpy as np

transcriptsv1 = pd.read_csv("transcripts_dataset_v1.csv.gz", compression="gzip")
transcriptsv2 = pd.read_csv("transcripts_dataset_v2.csv.gz", compression="gzip")
metadata = pd.read_csv("metadata.csv.gz", compression='gzip')

print('Transcipt v1 shape:', transcriptsv1.shape)
print('Transcipt v2 shape:', transcriptsv2.shape)
print('Metadata shape:', metadata.shape, '\n')

# print(list(transcriptsv1.columns))
# print(list(transcriptsv2.columns))
# print(list(metadata.columns))

# print('\n')

# print('Transcript v1 null check:\n', transcriptsv1.isna().sum(), '\n')
# print('Transcript v2 null check:\n', transcriptsv2.isna().sum(), '\n')
print('Metadata null check:\n', metadata.isna().sum()) # show_name 2 nulls, episode_description 205 nulls, category 6000, pubdate 20030


# concatenate transcripts dataframes
transcripts = pd.concat([transcriptsv1, transcriptsv2])
print(transcripts.shape)

# subset metadata
metadata = metadata[["show_name", "show_description", "publisher", "language", "episode_name", "episode_description", "duration", "show_filename_prefix", "episode_filename_prefix", 'category', 'pubdate']]
metadata = metadata.rename({"episode_filename_prefix": "episode_id", "show_filename_prefix": "show_id"}, axis="columns")  # rename cols

# remove episode_id suffix 
transcripts.episode_id = transcripts.episode_id.apply(lambda x: x.replace(".json", ""))
transcripts = transcripts.drop("Unnamed: 0", axis=1)

# remove whitespace
metadata.episode_id = metadata.episode_id.apply(lambda x: x.strip())

# join data
full_dataset = transcripts.join(metadata.set_index("episode_id"), on="episode_id", rsuffix="_trans")
full_dataset.isna().sum()

print(full_dataset.shape)
# full_dataset.head(5)

# Null inspection
full_dataset[full_dataset.isnull().any(axis=1)] # metadata discrepancies, episode_id present for all instances

### Correct mistake in word count from directory walk

# word_count was actually char count 
full_dataset = full_dataset.rename(columns={"word_count": "char_count"})

# count words
full_dataset["word_count"] = full_dataset.transcript.apply(lambda x: len(x.split(" ")))

languages_not_keep = ["['nl-BE']","['hi']","['id']","['ml']","['ms']","['pt']","['en-JM']","['en-IN']","['ga']","['ta']","['es']"]
languages = list(set(full_dataset.language.unique()) - set(languages_not_keep))

full_dataset = full_dataset[full_dataset['language'].isin(languages)]
print(full_dataset.shape)
full_dataset.language.value_counts()

# Save the DataFrame as a gzip-compressed CSV file
# full_dataset.to_csv('transcripts_dataset_final.csv.gz', compression='gzip', index=False)

## Filter out Education and Sports

transcripts_full = pd.read_csv("transcripts_dataset_final.csv.gz", compression="gzip")
print(transcripts_full.columns)
transcripts_full.shape

education_transcripts = transcripts_full[transcripts_full.category == 'Education']
education_transcripts.shape
# education_transcripts.to_csv('education_transcripts.csv.gz', compression='gzip')

sports_transcripts = transcripts_full[transcripts_full.category == 'Sports']
sports_transcripts.shape
# sports_transcripts.to_csv('sports_transcripts.csv.gz', compression='gzip')
