==> 3.get_probabilities.ipynb <==
# First step in pipeline
# Get probabilities of podcasts which you want to find segments and topics of
**Load HDBSCAN and t-SNE models and find probabilities for all sentences in a document. The output of this file will be used in the segmentation**

import os
import pickle
import hdbscan
import numpy as np
import pandas as pd
from openTSNE import TSNE
from bertopic import BERTopic
from sklearn.decomposition import PCA
from sentence_transformers import SentenceTransformer

# load PCA class object
with open('../Thesis/Final_models/pca_model_word100.pkl', 'rb') as inp:
    PCA_model = pickle.load(inp)

1-sum(PCA_model.explained_variance_ratio_)

# load t-SNE class object
with open('../Thesis/Final_models/tsne_data_word100.pkl', 'rb') as inp:
    tsne_data = pickle.load(inp)

# loop through all of the annotated transcripts and save dim reduced vectors
directory = '../Thesis/annotated_transcripts_input'

for dirpath, _, files in os.walk(directory):
    for file in files:
        episode_id = file.split('_')[0]
        path = os.path.join(dirpath, file)
        with open(path, errors='replace') as f: 
            prediction_documents = pd.read_csv(f)
            docs = prediction_documents.transcript_subset

            # embedd
            sentence_model = SentenceTransformer("all-MiniLM-L6-v2")
            embedded_sentences = sentence_model.encode(docs)

            # reduce dimensions with PCA and t-SNE
            PCA_data = PCA_model.transform(embedded_sentences)
            print(f"{1-sum(PCA_model.explained_variance_ratio_):.2%} of the variance has been removed by PCA for transcript {episode_id}")
            tsne_prediction = tsne_data.transform(PCA_data)

            # Write transcripts to files
            save_path = '../Thesis/annotated_dimreduced/'
            name_of_file = f'dimreduced_word100_{episode_id}.npy'
            complete_path = os.path.join(save_path, name_of_file)
            np.save(complete_path, tsne_prediction)

# load cluster model
with open('../Thesis/Final_models/hdbscan_msize15.pkl', 'rb') as inp:
    hdbscan_model = pickle.load(inp)

# loop through all of the annotated dim reduced embeddings and save topic probability density vector
dir = '../Thesis/annotated_dimreduced'

for dirpath, _, files in os.walk(dir):
    for file in files:
        episode_id = file[:-4].split('_')[2]
        path = os.path.join(dirpath, file)
        tsne_prediction = np.load(path)

        # get probability vectors for each cluster via soft clustering
        print(f'Starting {episode_id}')
        probabilities = hdbscan.membership_vector(hdbscan_model, tsne_prediction)
        print(f'nrows: {len(probabilities)}\nncols: {len(probabilities[0])}')
        # remedy HDBSCAN problem
        inds = np.where(np.isnan(probabilities))
        probabilities[inds] = 0
        print(f"Number of nan rows: {len(set(list(inds[:][0])))} for {episode_id}\n\n")

        # save probabilities to use in downstream segmentation
        name_of_file = f'../Thesis/annotated_probabilities_msize15/proba_dens_vec_msize15_{episode_id}'
        np.save(name_of_file, probabilities)

### Code for individual transcript predicitons

# load dataframe of one transcript with sentences as instances
sentences = pd.read_csv('first_podcast.csv.gz', compression='gzip')
documents = sentences.transcript_subset.to_list()

# embedd
sentence_model = SentenceTransformer("all-MiniLM-L6-v2")
embedded_sentences = sentence_model.encode(documents)

# reduce dimensions with PCA and t-SNE
PCA_model = PCA(n_components = 50)
PCA_data = PCA_model.fit_transform(embedded_sentences)
print(f"{1-sum(PCA_model.explained_variance_ratio_):.2%} of the variance has been removed by PCA")

tsne_test = tsne_data.transform(PCA_data)

# get probability vectors for each cluster via soft clustering
probabilities = hdbscan.membership_vector(hdbscan_model, tsne_test)
print(f'nrows: {len(probabilities)}')
print(f'ncols: {len(probabilities[0])}')
# remedy for HDBSCAN problem
inds = np.where(np.isnan(probabilities))
probabilities[inds] = 0
print(f"Number of nan rows: {len(set(list(inds[:][0])))}")

# save probabilities to use in segmentation
np.save('probabilities_new.npy', probabilities)
